Updated gadget control panel
Removed the old charm bracelet gadget because it doesn't make any sense in the world of new, larger, position retaining hooks.

Split gadget_menu out from master to embody the control panel. This panel now appears when the gadget is held and disappears when it's dropped.
