+++ websrc/webpack.config.js
@@ -95,7 +95,7 @@ module.exports =
 	createConfig( 'aardvark_monitor', 'Monitor', 'tsx' ),
 	createConfig( 'aardvark_renderer', 'Renderer', 'ts' ),
 	createConfig( 'test_panel', 'Test Panel', 'tsx' ),
-	createConfig( 'charm_bracelet', 'Charm Bracelet', 'tsx' ),
+	createConfig( 'gadget_menu', 'Gadget Menu', 'tsx' ),
 	createConfig( 'control_test', 'Control Tester', 'tsx' ),
 	{
 		target: "node",
